requirements-dev.txt
GitPython
requests
gradio
pywebview
pywin32-ctypes
pywin32
pyinstaller